# Optimized away INode.GetAllElements() allocations