--- a/src/AngleSharp/Dom/Internal/CollectionExtensions.cs
+++ b/src/AngleSharp/Dom/Internal/CollectionExtensions.cs
@@ -3,6 +3,8 @@
     using AngleSharp.Text;
     using System;
     using System.Collections.Generic;
+    using System.Collections;
+    using System.Linq;
 
     /// <summary>
     /// A bunch of methods for getting DOM elements.
@@ -158,22 +160,9 @@
 
         private static IEnumerable<T> GetAllElements<T>(this INode parent, Predicate<T> predicate)
             where T : class, INode
-        {
-            for (var i = 0; i < parent.ChildNodes.Length; i++)
-            {
-                var child = parent.ChildNodes[i] as T;
-
-                if (child != null && predicate(child))
-                {
-                    yield return child;
-                }
-
-                foreach (var element in parent.ChildNodes[i].GetAllElements<T>(predicate))
-                {
-                    yield return element;
-                }
-            }
-        }
+            => new NodeEnumerable(parent)
+            .OfType<T>()
+            .Where(node => predicate(node));
 
         private static IEnumerable<T> GetDescendendElements<T>(this INode parent, Predicate<T> predicate)
             where T : class, INode
@@ -187,6 +176,67 @@
                     yield return child;
                 }
             }
+        }
+
+        private class NodeEnumerable : IEnumerable<INode>
+        {
+            private readonly INode _startingNode;
+
+            public NodeEnumerable(INode startingNode)
+            {
+                _startingNode = startingNode;
+            }
+
+            public IEnumerator<INode> GetEnumerator() => new NodeEnumerator(_startingNode);
+            IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
+        }
+
+        private class NodeEnumerator : IEnumerator<INode>
+        {
+            private readonly Stack<EnumerationFrame> _frameStack = new Stack<EnumerationFrame>();
+
+            public NodeEnumerator(INode startingNode)
+            {
+                TryPushFrame(startingNode, 0);
+            }
+
+            public INode Current { get; private set; }
+            object IEnumerator.Current => Current;
+
+            public bool MoveNext()
+            {
+                if (_frameStack.Count == 0)
+                    return false;
+
+                var currentFrame = _frameStack.Pop();
+                Current = currentFrame.Parent.ChildNodes[currentFrame.ChildIndex];
+
+                TryPushFrame(currentFrame.Parent, currentFrame.ChildIndex + 1);
+                TryPushFrame(Current, 0);
+
+                return true;
+            }
+
+            private void TryPushFrame(INode parent, int childIndex)
+            {
+                if (childIndex < parent.ChildNodes.Length)
+                    _frameStack.Push(new EnumerationFrame(parent, childIndex));
+            }
+
+            public void Dispose() { }
+            public void Reset() { throw new NotSupportedException(); }
+        }
+
+        private struct EnumerationFrame
+        {
+            public EnumerationFrame(INode parent, int childIndex)
+            {
+                Parent = parent;
+                ChildIndex = childIndex;
+            }
+
+            public INode Parent { get; }
+            public int ChildIndex { get; }
         }
     }
 }