--- a/src/AngleSharp/Dom/Internal/HtmlCollection.cs
+++ b/src/AngleSharp/Dom/Internal/HtmlCollection.cs
@@ -26,7 +26,7 @@
             _elements = elements;
         }
 
-        public HtmlCollection(INode parent, Boolean deep = true, Predicate<T> predicate = null)
+        public HtmlCollection(INode parent, Boolean deep = true, Func<T, bool> predicate = null)
         {
             _elements = parent.GetElements(deep, predicate);
         }