@FlorianRappl 2016-12-01
Great change 👍 ! Memory is very important and this one is definitely a step in the right direction. Did the performance of large queries also improve?