@FlorianRappl 2016-12-02
Very cool - thanks. After PR I will adjust the style to adhere to the project (here single statements are not blocked + keywords instead of types).