@HellBrick 2016-12-02
>Did the performance of large queries also improve?

I haven't really tested large queries. This PR affects somewhat specific things (`IDocument`'s `Forms`, `Images`, `Scripts`, etc.) and no matter what you do from there, as long as you enumerate through all the results (without bailing early by `FirstOrDefault` or something), you had to pay the allocation price. The enumeration itself became a bit faster, but it was fast enough anyway, the whole point of the change is to avoid death by GC under load =)

Since I happen to have the numbers, here's the small speed benchmark:

```C#
public class TreeTraversingBenchmark
{
	private static readonly IDocument _document = new HtmlParser().ParseDocument( TestData.FastTorrent );

	[Benchmark]
	public void EnumerateAllNodes()
	{
		foreach ( var node in _document.All )
		{
		}
	}

	[Benchmark]
	public void EnumerateAllLinks()
	{
		foreach ( var node in _document.Links )
		{
		}
	}

	[Benchmark]
	public void EnumerateLinksWithImagesInside()
	{
		foreach ( var node in _document.Links.Where( l => l.FirstChild?.NodeName == "img" ) )
		{
		}
	}
}
```

```ini
BenchmarkDotNet=v0.10.0
OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i7 CPU 960 3.20GHz, ProcessorCount=8
Frequency=3162609 Hz, Resolution=316.1946 ns, Timer=TSC
Host Runtime=Clr 4.0.30319.42000, Arch=64-bit RELEASE [RyuJIT]
GC=Concurrent Workstation
JitModules=clrjit-v4.6.1586.0
Job Runtime(s):
        Clr 4.0.30319.42000, Arch=64-bit RELEASE [RyuJIT]
```

Before:
```
                         Method |        Mean |     StdDev |      Median |
------------------------------- |------------ |----------- |------------ |
              EnumerateAllNodes | 563.1366 us |  9.6620 us | 562.6725 us |
              EnumerateAllLinks | 499.8031 us | 26.8478 us | 493.6902 us |
 EnumerateLinksWithImagesInside | 509.4505 us |  8.0875 us | 506.4249 us |
```

After:
```
                         Method |        Mean |     StdDev |      Median |
------------------------------- |------------ |----------- |------------ |
              EnumerateAllNodes | 322.8301 us |  3.9090 us | 322.2199 us |
              EnumerateAllLinks | 405.7342 us | 12.3445 us | 401.5507 us |
 EnumerateLinksWithImagesInside | 396.6019 us |  1.9667 us | 396.3261 us |
```

>After PR I will adjust the style to adhere to the project (here single statements are not blocked + keywords instead of types)

Great, I'll look at the commit where you fix the code style and will try to follow the style better. Hopefully, this is not my last PR here, I have a few more optimizations in mind ;)