@FlorianRappl 2016-12-02
This is great news! Much appreciated 👍 

Thanks for the benchmark. I think the improvement is great - besides the max. 40% gain we also see that the standard deviation decreased a lot; so less allocations means less dependent on the system and much more stable performance. 🎉